moves hook signal in separate module
By moving signals in their own module, we can now follow Django guidelines and register signals on app.ready.
