+++ django_q/apps.py
@@ -6,3 +6,6 @@ from .conf import Conf
 class DjangoQConfig(AppConfig):
     name = 'django_q'
     verbose_name = Conf.LABEL
+
+    def ready(self):
+        from django_q.signals import call_hook
